Update node.py
Change of node variable channels from "list" to "array". At initialisation an array of the length of maximum number of possible channels is generated using the "get_capabilities" funciton
Introduction of the function "get_capabilities" by using the "Message.ID.RESPONSE_CAPABILITIES" to get the capabilities of the ant+ chip.